fixed awful documentation and code cleanup
