+++ crates/distances/src/strings/needleman_wunsch/helpers.rs
@@ -28,13 +28,17 @@ pub enum Edit {
 
 /// Computes the Needleman-Wunsch dynamic programming table for two sequences.
 ///
-/// The penalty for a mismatch is 1, and the penalty for a gap is 1. There is no
-/// penalty for a match. Our implementation minimizes the total penalty.
+/// Users can input custom penalities, but those penalties should satisfy the following:
+/// * All penalties are non-negative
+/// * The match "penalty" is 0
+///
+///  Our implementation *minimizes* the total penalty.
 ///
 /// # Arguments
 ///
 /// * `x`: The first sequence.
 /// * `y`: The second sequence.
+/// * `penalties`: The penalties for a match, mismatch, and gap.
 ///
 /// # Returns
 ///
@@ -92,85 +96,6 @@ fn min2<U: UInt>(a: (U, Direction), b: (U, Direction)) -> (U, Direction) {
         b
     }
 }
-
-/// Converts two aligned sequences into vectors of edits.
-///
-/// # Arguments
-///
-/// * `x`: The first aligned sequence.
-/// * `y`: The second aligned sequence.
-///
-/// # Returns
-///
-/// A 2-slice of Vec<Edit>, each containing the edits needed to convert one
-/// sequence into the other (substitutions only since they are both aligned).
-#[must_use]
-pub fn compute_edits(x: &str, y: &str) -> [Vec<Edit>; 2] {
-    [_x_to_y(x, y), _x_to_y(y, x)]
-}
-
-/// Helper for `compute_edits` to compute the edits for turning aligned `x` into aligned `y`.
-///
-/// Expects `x` and `y` to be aligned sequences generated by `trace_back_iterative` or `trace_back_recursive`.
-/// Returns a vector of substitutions needed to turn `x` into `y`.
-///
-/// # Arguments
-///
-/// * `x`: The first aligned sequence.
-/// * `y`: The second aligned sequence.
-///
-/// # Returns
-///
-/// A vector of edits needed to convert `x` into `y`.
-#[must_use]
-pub fn _x_to_y(x: &str, y: &str) -> Vec<Edit> {
-    x.chars()
-        .zip(y.chars())
-        .enumerate()
-        .filter(|(_, (x, y))| x != y)
-        .map(|(i, (_, y))| Edit::Sub(i, y))
-        .collect()
-}
-
-/// Given two aligned strings, returns into a sequence of edits to transform the unaligned version of one into the unaligned version of the other.
-///
-/// Requires that gaps are never aligned with gaps (our NW implementation, with the default penalties ensures this).
-/// Expects to receive two strings which were aligned using either `trace_back_iterative` or `trace_back_recursive`.
-///
-/// # Arguments
-///
-/// * `x`: An aligned string.
-/// * `y`: An aligned string.
-///
-/// # Returns
-///
-/// A vector of edits to transform the unaligned version of `x` into the unaligned version of `y`.
-#[must_use]
-pub fn unaligned_x_to_y(x: &str, y: &str) -> Vec<Edit> {
-    let mut unaligned_x_to_y = Vec::new();
-    let mut modifier = 0;
-
-    x.chars()
-        .zip(y.chars())
-        .enumerate()
-        .filter(|(_, (x, y))| x != y)
-        .for_each(|(index, (c_x, c_y))| {
-            let i = index - modifier;
-
-            if c_x == '-' {
-                unaligned_x_to_y.push(Edit::Ins(i, c_y));
-            } else if c_y == '-' {
-                unaligned_x_to_y.push(Edit::Del(i));
-                modifier += 1;
-            } else {
-                unaligned_x_to_y.push(Edit::Sub(i, c_y));
-            }
-        });
-
-    unaligned_x_to_y
-}
-
-// pub fn subs_to_general_edits(x: )
 
 /// Iteratively traces back through the Needleman-Wunsch table to get the alignment of two sequences.
 ///
@@ -305,16 +230,95 @@ fn _trace_back_recursive<U: UInt>(
     }
 }
 
+/// Converts two aligned sequences into vectors of edits.
+///
+/// # Arguments
+///
+/// * `x`: The first aligned sequence.
+/// * `y`: The second aligned sequence.
+///
+/// # Returns
+///
+/// A 2-slice of Vec<Edit>, each containing the edits needed to convert one aligned
+/// sequence into the other.
+/// Since both input sequences are aligned, all edits are substitutions in the returned vectors are Substitutions.
+#[must_use]
+pub fn compute_edits(x: &str, y: &str) -> [Vec<Edit>; 2] {
+    [_x_to_y(x, y), _x_to_y(y, x)]
+}
+
+/// Helper for `compute_edits` to compute the edits for turning aligned `x` into aligned `y`.
+///
+/// Expects `x` and `y` to be aligned sequences generated by `trace_back_iterative` or `trace_back_recursive`.
+/// Returns a vector of substitutions needed to turn `x` into `y`.
+///
+/// # Arguments
+///
+/// * `x`: The first aligned sequence.
+/// * `y`: The second aligned sequence.
+///
+/// # Returns
+///
+/// A vector of edits needed to convert `x` into `y`.
+#[must_use]
+pub fn _x_to_y(x: &str, y: &str) -> Vec<Edit> {
+    x.chars()
+        .zip(y.chars())
+        .enumerate()
+        .filter(|(_, (x, y))| x != y)
+        .map(|(i, (_, y))| Edit::Sub(i, y))
+        .collect()
+}
+
+/// Given two aligned strings, returns into a sequence of edits to transform the unaligned
+/// version of one into the unaligned version of the other.
+///
+/// Requires that gaps are never aligned with gaps (our NW implementation with the default penalties ensures this).
+/// Expects to receive two strings which were aligned using either `trace_back_iterative` or `trace_back_recursive`.
+///
+/// # Arguments
+///
+/// * `x`: An aligned string.
+/// * `y`: An aligned string.
+///
+/// # Returns
+///
+/// A vector of edits to transform the unaligned version of `x` into the unaligned version of `y`.
+#[must_use]
+pub fn unaligned_x_to_y(x: &str, y: &str) -> Vec<Edit> {
+    let mut unaligned_x_to_y = Vec::new();
+    let mut modifier = 0;
+
+    x.chars()
+        .zip(y.chars())
+        .enumerate()
+        .filter(|(_, (x, y))| x != y)
+        .for_each(|(index, (c_x, c_y))| {
+            let i = index - modifier;
+
+            if c_x == '-' {
+                unaligned_x_to_y.push(Edit::Ins(i, c_y));
+            } else if c_y == '-' {
+                unaligned_x_to_y.push(Edit::Del(i));
+                modifier += 1;
+            } else {
+                unaligned_x_to_y.push(Edit::Sub(i, c_y));
+            }
+        });
+
+    unaligned_x_to_y
+}
+
 /// Applies a set of edits to a reference (unaligned) string to get a target (unaligned) string.
 ///
 /// # Arguments
 ///
-/// * `x`: The reference string.
+/// * `x`: The unaligned reference string.
 /// * `edits`: The edits to apply to the reference string.
 ///
 /// # Returns
 ///
-/// The target string.
+/// The unaligned target string.
 #[must_use]
 pub fn apply_edits(x: &str, edits: &[Edit]) -> String {
     let mut x: Vec<char> = x.chars().collect();
@@ -394,22 +398,5 @@ mod tests {
         let (aligned_x, aligned_y) = trace_back_iterative(&guilty_table, [guilty_x, guilty_y]);
         assert_eq!(aligned_x, "NOTGUILTY");
         assert_eq!(aligned_y, "NOTGUILTY");
-    }
-
-    #[test]
-    fn test_apply_edits() {
-        let x = "ACBAAB";
-        let edits: Vec<Edit> = vec![
-            Edit::Ins(1, 'A'),
-            Edit::Sub(3, 'A'),
-            Edit::Del(4),
-            Edit::Sub(5, 'C'),
-        ];
-
-        let actual = apply_edits(x, &edits);
-
-        let expected = "AACAAC";
-
-        assert_eq!(actual, expected);
     }
 }
